gst/spectrum/gstspectrum.*: Fix type mixup in spectrum->interval (gdouble<->guint64). Spotted by
Original commit message from CVS:
* gst/spectrum/gstspectrum.c: (gst_spectrum_init),
(gst_spectrum_set_property):
* gst/spectrum/gstspectrum.h:
Fix type mixup in spectrum->interval (gdouble<->guint64). Spotted by
René Stadler

+++ gst/spectrum/gstspectrum.h
@@ -48,7 +48,7 @@ struct _GstSpectrum {
 
   /* properties */
   gboolean message;             /* whether or not to post messages */
-  gdouble interval;             /* how many seconds between emits */
+  guint64 interval;             /* how many seconds between emits */
   guint bands;                  /* number of spectrum bands */
   gint threshold;               /* energy level treshold */
